# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ardi - Autonomous Penetration Testing Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update the current user's profile and notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Dashboard
...
  /assets:
    get:
      operationId: getAssets
      tags: [assets]
      summary: List all assets
      responses:
        "200":
          description: Asset list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create a new asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get asset by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Update asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
  /findings/{id}:
    get:
      operationId: getFinding
      tags: [findings]
      summary: Get finding by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Finding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
    patch:
      operationId: updateFinding
      tags: [findings]
      summary: Update finding status or notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          
        password:
          type: string
    RegisterInput:
      type: object
      required: [email, password, name, orgName]
      properties:
        email:
          type: string
          
        password:
          type: string
        name:
          type: string
        orgName:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, name, role, orgName, notifyScanComplete, notifyCritical, notifyWeeklyDigest]
      properties:
        id:
          type: number
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, analyst, viewer]
        orgName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        notifyScanComplete:
          type: boolean
        notifyCritical:
          type: boolean
        notifyWeeklyDigest:
          type: boolean
        createdAt:
          type: string
...
    UpdateMeInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        orgName:
          type: string
          minLength: 1
        notifyScanComplete:
          type: boolean
        notifyCritical:
          type: boolean
        notifyWeeklyDigest:
          type: boolean

    # Dashboard
...
    Asset:
      type: object
      required: [id, name, type, target, status, riskLevel, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        type:
          type: string
          enum: [web_app, network, cloud_aws, cloud_azure, cloud_gcp, api, mobile]
        target:
          type: string
        status:
          type: string
          enum: [active, inactive, scanning, pending]
        riskLevel:
          type: string
          enum: [critical, high, medium, low, none]
        tags:
          type: array
          items:
            type: string
        lastScannedAt:
          type: ["string", "null"]
        openFindings:
          type: number
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    AssetInput:
      type: object
      required: [name, type, target]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [web_app, network, cloud_aws, cloud_azure, cloud_gcp, api, mobile]
        target:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
    AssetUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, inactive, scanning, pending]
        description:
          type: string
        tags:
          type: array
          items:
            type: string
...
    Finding:
      type: object
      required: [id, title, severity, status, category, assetId, assetName, scanId, cve, mitreId, mitreTactic, mitreTechnique, description, remediation, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        status:
          type: string
          enum: [open, in_progress, resolved, accepted_risk, false_positive]
        category:
          type: string
          # Web-application categories, plus the infrastructure categories the
          # nmap and nuclei engines actually emit (network, tls, dns, exposure,
          # web, file, other). The engine is the source of truth here: a
          # category it can produce but the spec rejects makes the whole
          # findings endpoint fail Zod validation and return 500.
          enum: [injection, broken_auth, xss, idor, misconfiguration, exposed_data, privilege_escalation, lateral_movement, persistence, exfiltration, network, web, tls, dns, exposure, file, other]
        assetId:
          type: number
        assetName:
          type: string
        scanId:
          type: number
        cve:
          type: ["string", "null"]
        cvss:
          type: ["number", "null"]
        mitreId:
          type: string
        mitreTactic:
          type: string
        mitreTechnique:
          type: string
        description:
          type: string
        remediation:
          type: string
        evidence:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    FindingUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, resolved, accepted_risk, false_positive]
        notes:
          type: string
...