# lib/api-zod/src/generated/types/updateMeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ardi - Autonomous Penetration Testing Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMeInput {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  orgName?: string;
  notifyScanComplete?: boolean;
  notifyCritical?: boolean;
  notifyWeeklyDigest?: boolean;
}