# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ardi - Autonomous Penetration Testing Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  orgName: string;
  /** @nullable */
  avatarUrl?: string | null;
  notifyScanComplete: boolean;
  notifyCritical: boolean;
  notifyWeeklyDigest: boolean;
  createdAt?: string;
}